back_end/pyproject.toml
[tool.poetry]
name = "remember-back-end"
version = "0.1.0"
description = ""
authors = ["Madi <madi_gaziz@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2"
django-rest-framework = "^0.1.0"
black = "^25.1.0"
aiogram = "^3.20.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.1"
requests = "^2.32.3"
pip = "^25.0.1"
python-dotenv = "^1.1.0"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
